+++ src/main/java/com/dabomstew/pkrandom/romhandlers/Gen5RomHandler.java
@@ -1166,6 +1166,7 @@ public class Gen5RomHandler extends AbstractDSRomHandler {
             getTextHandler().bw1NacreneCityTextModifications(taggedTypes, this.random);
             getTextHandler().bw1CasteliaCityPraiseTextModifications(taggedTypes);
             getTextHandler().bw1CasteliaCityBurghTextModifications(taggedTypes);
+            getTextHandler().bw1CherenBurghTextModifications(taggedTypes);
         }
     }